- Strategy pattern implemented - Create strategy pattern under model folder - 2 strategies were added. - 1. First past the post / traditional method - 2. Weighted average method

+++ src/main/java/com/votingsystem/VotingSystem/model/Option.java
@@ -31,10 +31,12 @@ public class Option {
     private int voteCount;
 
     @Column(name = "VotePercentage")
-    private int votePercentage;
+    private double votePercentage;
 
     @ManyToOne
     @JoinColumn(name = "PollID", nullable = false)
     private Poll poll;
+
+	 
 
 }
